+++ src/__mocks__/resolve-cloud-common/postgres.ts
@@ -1,0 +1,3 @@
+export const executeStatement = jest.fn()
+
+export const escapeId = (str: string) => str