+++ src/__mocks__/resolve-cloud-common/utils.ts
@@ -1,0 +1,2 @@
+export const log = jest.fn()
+export const retry = jest.fn()